[wgsl] Update `static_assert` to `const_assert`.
The name `static_assert` has been updated in the spec to `const_assert`.
This CL updates the CTS to match.

Fixes #2178
