+++ src/webgpu/shader/validation/const_assert/const_assert.spec.ts
@@ -1,4 +1,4 @@
-export const description = `Validation tests for static_assert`;
+export const description = `Validation tests for const_assert`;
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { ShaderValidationTest } from '../shader_validation_test.js';
@@ -6,14 +6,14 @@ import { ShaderValidationTest } from '../shader_validation_test.js';
 export const g = makeTestGroup(ShaderValidationTest);
 
 /**
- * Builds a static_assert() statement, which checks that @p expr is equal to @p expect_true.
+ * Builds a const_assert() statement, which checks that @p expr is equal to @p expect_true.
  * @param expect_true true if @p expr should evaluate to true
  * @param expr the constant expression
  * @param scope module-scope or function-scope constant expression
  * @returns the WGSL code
  */
 function buildStaticAssert(expect_true: boolean, expr: string, scope: 'module' | 'function') {
-  const stmt = expect_true ? `static_assert ${expr};` : `static_assert !(${expr});`;
+  const stmt = expect_true ? `const_assert ${expr};` : `const_assert !(${expr});`;
   return scope === 'module' ? stmt : `fn f() { ${stmt} }`;
 }
 
@@ -28,7 +28,7 @@ const kConditionCases = {
 };
 
 g.test('constant_expression')
-  .desc(`Test that static_assert validates the condition expression.`)
+  .desc(`Test that const_assert validates the condition expression.`)
   .params(u =>
     u
       .combine('case', Object.keys(kConditionCases) as Array<keyof typeof kConditionCases>)
@@ -47,7 +47,7 @@ const three = 2;
   });
 
 g.test('evaluation_stage')
-  .desc(`Test that the static_assert expression must be a constant expression.`)
+  .desc(`Test that the const_assert expression must be a constant expression.`)
   .params(u =>
     u
       .combine('scope', ['module', 'function'] as const)
